+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 1.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Fix permission Check.
+  [ebrehault]
 
 1.0.1 (2023-10-03)
 ------------------